Fix stateful runtime CI fixture
